refactor: extract fifty fifty stats accumulator
